requirements.txt
autopep8==1.4.4
awscli==1.16.287
boto3==1.10.23
botocore==1.13.23
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
docutils==0.15.2
Flask==1.1.1
future==0.18.1
gunicorn==19.9.0
idna==2.8
imgurpython==1.1.7
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
line-bot-sdk==1.14.0
MarkupSafe==1.1.1
Pillow==6.2.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.8
pycodestyle==2.5.0
python-dateutil==2.8.0
PyYAML==5.1.2
qrcode==6.1
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
six==1.13.0
urllib3==1.25.6
Werkzeug==0.16.0